Added drand48(3), lrand48(3), mrand48(3).
These had been lying around for a while, partially because they aren't
very interesting, but mostly because the implementations from eglibc,
dietlibc and musl libc (with the same seeds) give different results.

This is ridiculous, but it also happens to have a very low priority on
my todo list.

The mtime is recent because I modified some comments.