Fixed a typo, tested 1e6th element before getting straight to 1e9
Reason: I converted the C version to output one element at a time rather
than using refill and introduced an off-by-one which happened everytime
the index had to start over at zero (ie once every 40k random numbers).

The idea is to leave this file here for reference (as it's a faithful
translation of the original code) and switch to the (trivially) reworked
version.