Missed an invalid seed in marsaglia.vim; corrected the wikipedia entry
The funny thing is that even Marsaglia forgets to mention them in his
posts, only warning to avoid zero :)
Anyway, the problem is that the algorithm used is a 16-bit MWC in
disguise, that is, m_w (and m_z) is actually a compact encoding for
MWC {
    x[r];
    carry
}

with r=1 and all numbers being 16 bits wide; rather than keeping two
pesky uint16 around, do the computations using 32 bits math, going back
to 16 bit and starting all over again *every* *single* *time*,
prof Marsaglia decided he'd encode those numbers as the high/low part
of a 32 bit number (which is mathematically sound btw).
The problem is that a mwc has **two** fixed points, one being all
zeroes, and the other being when all x's equal b-1 (here: 0xFFFF)
and the carry is a-1 (the multiplier -- that's why the numbers are in
hexadecimal).

Shame on me for not noticing before!