Fixed IA initialization; IA ok on non-32 bits
IA only needs the bits in a word to be >= 2*ALPHA, which is ok
because ALPHA==8 by default and vim guarantees when it is compiled
to pick an integer type that's >= 32 bits wide.

My old misinterpretation that we needed 2*ALPHA bits **set** in the
internal state stuck for so long because it didn't make sense - I
couldn't see where on earth it came from, and it puzzled me, so I failed
to notice that it just wasn't there.
The fact that IA sucks at getting out of "zeroland" made the
misunderstanding look more plausible, as I was focused on IA vs IBAA
("ok, IBAA fails with all zero, but why does IA need 16 bits instead?")

The idea for the proof was suggested by reading the ranrot paper
available at http://www.agner.org/random/theory/chaosran.pdf;
my previous idea was much more complicated.

The naming "IA()", "s:ia()" is confusing.