Added NSA's SPECK cipher.
Because googling "vim+NSA" or similar gives irrelevant results.
And because I haven't tried assembly, but it seems that optimizing
C compilers can do **wonders** on this cipher, but the magic gets lost
as you go to more abstract languages - eg in java,C# SPECK's performance
decreases faster than that of other ciphers.
So I just __had__ to try it in vimscript!